Deprecated and Defunct functions in ‘PowerUpR’

Deprecated functions are provided for compatibility purposes and will be defunct in the next ‘PowerUpR’ release. Defunct functions are no longer supported.

Details

Deprecated and defunct functions and their replacement are as follows:

mrss.ira1r1 is deprecated, use cosa.ira1r1

mrss.cra2r2 is deprecated, use cosa.cra2r2

mrss.cra3r3 is deprecated, use cosa.cra3r3

mrss.cra4r4 is deprecated, use cosa.cra4r4

mrss.bcra3f2 is deprecated, use cosa.bcra3f2

mrss.bcra3r2 is deprecated, use cosa.bcra3r2

mrss.bcra4r2 is deprecated, use cosa.bcra4r2

mrss.bcra4f3 is deprecated, use cosa.bcra4f3

mrss.bcra4r2 is deprecated, use cosa.bcra4r2

mrss.bira4r1 is deprecated, use cosa.bira4r1

mrss.bira3r1 is deprecated, use cosa.bira3r1

mrss.bira2c1 is deprecated, use cosa.bira2c1

mrss.bira2f1 is deprecated, use cosa.bira2f1

mrss.bira2r1 is deprecated, use cosa.bira2r1

mrss.to.mdes is defunct, use cosa.to.mdes

mrss.to.power is defunct, use cosa.to.power

optimal.ira1r1 is deprecated, use cosa.ira1r1

optimal.cra2r2 is deprecated, use cosa.cra2r2

optimal.cra3r3 is deprecated, use cosa.cra3r3

optimal.cra4r4 is deprecated, use cosa.cra4r4

optimal.bcra3f2 is deprecated, use cosa.bcra3f2

optimal.bcra3r2 is deprecated, use cosa.bcra3r2

optimal.bcra4r2 is deprecated, use cosa.bcra4r2

optimal.bcra4f3 is deprecated, use cosa.bcra4f3

optimal.bcra4r2 is deprecated, use cosa.bcra4r2

optimal.bira4r1 is deprecated, use cosa.bira4r1

optimal.bira3r1 is deprecated, use cosa.bira3r1

optimal.bira2c1 is deprecated, use cosa.bira2c1

optimal.bira2f1 is deprecated, use cosa.bira2f1

optimal.bira2r1 is deprecated, use cosa.bira2r1

optimal.to.mdes is deprecated, use cosa.to.mdes

optimal.to.power is deprecated, use cosa.to.power

plot.pars is defunct, there is no replacement function

t1t2.error is defunct, there is no replacement function

Format

Deprecated functions are no longer documented, please use documentation for new functions.