COSA Solver for Moderator Effect in Three-Level Cluster Random Assignment Designs, Treatment and Moderator at Level-3

cosa.mod3.cra3r3 finds constrained optimal sample allocation (COSA) solutions for moderator effect in designs with three-levels where level-3 units are randomly assigned to treatment and control groups and moderator is at level-3. COSA can be found in the following forms, (i) under budgetary constraints given marginal costs per unit, (ii) under power contraints given marginal costs per unit, (iii) under MDES contraints given marginal costs per unit, and (iv) under sample size contraints for one or more levels along with any of the i,ii, or iii options.

cosa.mod3.cra3r3(cn=0, cJ=0, cK=0, cost=NULL,
                 n=NULL, J=NULL, K=NULL, P=NULL, nJK0=c(10,10,10), P0=.50,
                                 constrain="power", optimizer="auglag_slsqp",
                                 gm=2, ncase=10,
                 power=.80, mdes=.25, alpha=.05, two.tail=TRUE,
                 rho2, rho3, Q=NULL,
                 g3=0, R12=0, R22=0, R32=0)

Arguments

cn
marginal cost per level-1 unit.

cJ
marginal cost per level-2 unit.

cK
marginal cost per level-3 unit.

cost
total cost or budget.

n
harmonic mean of level-1 units across level-2 units (or simple average).

J
harmonic mean of level-2 units across level-3 units (or simple average).

K
level-3 sample size.

P
proportion of level-3 units randomly assigned to treatment.

nJK0
vector with a length of two to specifiy starting values for level-1, level-2 and level-3 sample sizes.

P0
starting value for proportion of level-3 units randomly assigned to treatment.

constrain
one of the followings can be constrained at a specified cost or value: "cost", "power", or "mdes".

optimizer
algorithm to find optimal sample sizes given total cost, power, or MDES. Available algorithms: "auglag_cobyla", "auglag_lbfgs", "auglag_mma", or "auglag_slsqp".

gm
grid multiplier to increase the range of sample sizes for approximate integer solutions.

ncase
number of approximate integer solutions to be exported.

power
statistical power (1 - type II error).

mdes
minimum detectable effect size.

alpha
probability of type I error.

two.tail
logical; TRUE for two-tailed hypothesis testing, FALSE for one-tailed hypothesis testing.

rho2
proportion of variance in the outcome explained by level-2 units.

rho3
proportion of variance in the outcome explained by level-3 units.

Q
proportion of level-1 units within moderator subgroup. NULL implies continuous moderator.

g3
number of covariates at level-3 excluding moderator.

R12
proportion of level-1 variance in the outcome explained by level-1 covariates.

R22
proportion of level-2 variance in the outcome explained by level-2 covariates.

R32
proportion of level-3 variance in the outcome explained by level-3 covariates.

Value

fun

function name.

parms

list of parameters used in the function.

nloptr

list of nloptr log and output.

exact.optim

exact solution. MDES is calculated at the specified power (default .80), and power is calculated at the specified MDES (default .25).

round.optim

rounded solution. MDES is calculated at the specified power (default .80), and power is calculated at the specified MDES (default .25).

integer.optim

approximate integer solutions. MDES is calculated at the specified power (default .80), and power is calculated at the specified MDES (default .25).

See also

mdes.mod3.cra3r3, power.mod3.cra3r3

Examples

# NOT RUN {
      cosa.mod3.cra3r3(cn=1, cJ=10, cK=100, n=10, constrain="mdes",
                          rho2=.20, rho3=.10)


# }