MDES Calculator for Two-Level Fixed Effects Blocked Individual Random Assignment Designs, Treatment at Level-1

mdes.bira2f1 calculates minimum detectable effect size (MDES) for designs with two-levels where level-1 units are randomly assigned to treatment and control groups within level-2 units (fixed blocks).

mdes.bira2f1(power=.80, alpha=.05, two.tail=TRUE,
              P=.50, g1=0, R12=0,
              n, J, ...)

Arguments

power
statistical power \((1-\beta)\).

alpha
probability of type I error.

two.tail
logical; TRUE for two-tailed hypothesis testing, FALSE for one-tailed hypothesis testing.

P
average proportion of level-1 units randomly assigned to treatment within level-2 units.

g1
number of covariates at level-1.

R12
proportion of level-1 variance in the outcome explained by level-1 covariates.

n
harmonic mean of level-1 units across level-2 units (or simple average).

J
level-2 sample size.

...
to handle extra parameters passed from other functions, do not define any additional parameters.

Value

fun

function name.

parms

list of parameters used in MDES calculation.

ncp

noncentrality parameter.

mdes

minimum detectable effect size and \((1-\alpha)\)% lower and upper confidence limits.

See also

power.bira2f1, cosa.bira2f1

Examples

# NOT RUN {
 mdes.bira2f1(n=55, J=3)


# }